Renamed `database.py` to `player_assets.py`
