+++ website/database/player.py
@@ -1,5 +1,5 @@
 from website import db
-from website.database.database import (
+from website.database.player_assets import (
     Under_construction,
 )
