+++ website/views.py
@@ -10,7 +10,7 @@ from .database.messages import Chat
 
 from .database.player import Player
 from . import db
-from .database.database import Resource_on_sale
+from .database.player_assets import Resource_on_sale
 from .utils import check_existing_chats
 
 views = Blueprint("views", __name__)
